requirements.txt
torch==2.0.1
torchvision==0.15.2
scikit-learn==1.3.0
numpy==1.25.2
opencv-python==4.8.0.76
pillow==10.0.0
matplotlib==3.7.2
tqdm==4.66.1